+++ packages/interface/src/components/Explorer/views/MediaView/MediaViewItem.tsx
@@ -1,31 +1,9 @@
 import clsx from "clsx";
 import { memo } from "react";
-import {
-	Eye,
-	Copy,
-	Trash,
-	MagnifyingGlass,
-	Image,
-	Video,
-	Microphone,
-	FileText,
-	Stack,
-	Sparkle,
-	FilmStrip,
-	Waveform,
-	TextAa,
-	Crop,
-	FileVideo,
-} from "@phosphor-icons/react";
 import type { File } from "@sd/ts-client";
 import { File as FileComponent } from "../../File";
-import { useExplorer } from "../../context";
 import { useSelection } from "../../SelectionContext";
-import { getContentKind } from "../../utils";
-import { useContextMenu } from "../../../../hooks/useContextMenu";
-import { useJobDispatch } from "../../../../hooks/useJobDispatch";
-import { useLibraryMutation } from "../../../../context";
-import { usePlatform } from "../../../../platform";
+import { useFileContextMenu } from "../../hooks/useFileContextMenu";
 
 function formatDuration(seconds: number): string {
 	const mins = Math.floor(seconds / 60);
@@ -55,326 +33,12 @@ export const MediaViewItem = memo(function MediaViewItem({
 	onSelect,
 	size,
 }: MediaViewItemProps) {
-	const { currentPath } = useExplorer();
 	const { selectedFiles } = useSelection();
-	const platform = usePlatform();
-	const copyFiles = useLibraryMutation("files.copy");
-	const deleteFiles = useLibraryMutation("files.delete");
-	const { runJob } = useJobDispatch();
 
-	const getTargetFiles = () => {
-		if (selected && selectedFiles.length > 0) {
-			return selectedFiles;
-		}
-		return [file];
-	};
-
-	const contextMenu = useContextMenu({
-		items: [
-			{
-				icon: Eye,
-				label: "Quick Look",
-				onClick: () => {
-					console.log("Quick Look:", file.name);
-				},
-				keybind: "Space",
-			},
-			{
-				icon: MagnifyingGlass,
-				label: "Show in Finder",
-				onClick: async () => {
-					if ("Physical" in file.sd_path) {
-						const physicalPath = file.sd_path.Physical.path;
-						if (platform.revealFile) {
-							try {
-								await platform.revealFile(physicalPath);
-							} catch (err) {
-								console.error("Failed to reveal file:", err);
-								alert(`Failed to reveal file: ${err}`);
-							}
-						}
-					}
-				},
-				keybind: "⌘⇧R",
-				condition: () =>
-					"Physical" in file.sd_path && !!platform.revealFile,
-			},
-			{ type: "separator" },
-			{
-				icon: Copy,
-				label:
-					selected && selectedFiles.length > 1
-						? `Copy ${selectedFiles.length} items`
-						: "Copy",
-				onClick: async () => {
-					const targets = getTargetFiles();
-					const sdPaths = targets.map((f) => f.sd_path);
-
-					window.__SPACEDRIVE__ = window.__SPACEDRIVE__ || {};
-					window.__SPACEDRIVE__.clipboard = {
-						operation: "copy",
-						files: sdPaths,
-						sourcePath: currentPath,
-					};
-
-					console.log(`Copied ${sdPaths.length} files to clipboard`);
-				},
-				keybind: "⌘C",
-			},
-			{
-				icon: Copy,
-				label: "Paste",
-				onClick: async () => {
-					const clipboard = window.__SPACEDRIVE__?.clipboard;
-					if (!clipboard || !clipboard.files || !currentPath) {
-						return;
-					}
-
-					try {
-						await copyFiles.mutateAsync({
-							sources: { paths: clipboard.files },
-							destination: currentPath,
-							overwrite: false,
-							verify_checksum: false,
-							preserve_timestamps: true,
-							move_files: false,
-							copy_method: "Auto" as const,
-						});
-					} catch (err) {
-						console.error("Failed to paste:", err);
-					}
-				},
-				keybind: "⌘V",
-				condition: () => {
-					const clipboard = window.__SPACEDRIVE__?.clipboard;
-					return (
-						!!clipboard &&
-						!!clipboard.files &&
-						clipboard.files.length > 0
-					);
-				},
-			},
-			// Media Processing submenus
-			{
-				type: "submenu",
-				icon: Image,
-				label: "Image Processing",
-				condition: () => getContentKind(file) === "image",
-				submenu: [
-					{
-						icon: Sparkle,
-						label: "Generate Blurhash",
-						onClick: async () => {
-							const targets = getTargetFiles();
-							await runJob("thumbnail", {
-								file_ids: targets.map((f) => f.id),
-								generate_blurhash: true,
-							});
-						},
-						condition: () => !file.image_media_data?.blurhash,
-					},
-					{
-						icon: Crop,
-						label: "Regenerate Thumbnail",
-						onClick: async () => {
-							const targets = getTargetFiles();
-							await runJob("thumbnail", {
-								file_ids: targets.map((f) => f.id),
-								force: true,
-							});
-						},
-					},
-					{
-						icon: TextAa,
-						label: "Extract Text (OCR)",
-						onClick: async () => {
-							const targets = getTargetFiles();
-							await runJob("ocr", {
-								file_ids: targets.map((f) => f.id),
-							});
-						},
-						keybind: "⌘⇧T",
-					},
-				],
-			},
-			{
-				type: "submenu",
-				icon: Video,
-				label: "Video Processing",
-				condition: () => getContentKind(file) === "video",
-				submenu: [
-					{
-						icon: FilmStrip,
-						label: "Generate Thumbstrip",
-						onClick: async () => {
-							const targets = getTargetFiles();
-							await runJob("thumbstrip", {
-								file_ids: targets.map((f) => f.id),
-								frame_count: 10,
-							});
-						},
-						condition: () =>
-							!file.sidecars?.some(
-								(s) => s.kind === "thumbstrip",
-							),
-					},
-					{
-						icon: Sparkle,
-						label: "Generate Blurhash",
-						onClick: async () => {
-							const targets = getTargetFiles();
-							await runJob("thumbnail", {
-								file_ids: targets.map((f) => f.id),
-								generate_blurhash: true,
-							});
-						},
-						condition: () => !file.video_media_data?.blurhash,
-					},
-					{
-						icon: Crop,
-						label: "Regenerate Thumbnail",
-						onClick: async () => {
-							const targets = getTargetFiles();
-							await runJob("thumbnail", {
-								file_ids: targets.map((f) => f.id),
-								force: true,
-							});
-						},
-					},
-					{
-						icon: Waveform,
-						label: "Extract Subtitles",
-						onClick: async () => {
-							const targets = getTargetFiles();
-							await runJob("speech_to_text", {
-								file_ids: targets.map((f) => f.id),
-								output_format: "srt",
-							});
-						},
-					},
-					{
-						icon: FileVideo,
-						label: "Generate Proxy",
-						onClick: async () => {
-							const targets = getTargetFiles();
-							await runJob("proxy", {
-								file_ids: targets.map((f) => f.id),
-								quality: "720p",
-							});
-						},
-						keybind: "⌘⇧P",
-					},
-				],
-			},
-			{
-				type: "submenu",
-				icon: Microphone,
-				label: "Audio Processing",
-				condition: () => getContentKind(file) === "audio",
-				submenu: [
-					{
-						icon: TextAa,
-						label: "Transcribe Audio",
-						onClick: async () => {
-							const targets = getTargetFiles();
-							await runJob("speech_to_text", {
-								file_ids: targets.map((f) => f.id),
-								model: "whisper-base",
-							});
-						},
-						keybind: "⌘⇧T",
-					},
-				],
-			},
-			// Batch operations
-			{
-				type: "submenu",
-				icon: Stack,
-				label: `Process ${selectedFiles.length} Items`,
-				condition: () => selected && selectedFiles.length > 1,
-				submenu: [
-					{
-						icon: Crop,
-						label: "Regenerate All Thumbnails",
-						onClick: async () => {
-							await runJob("thumbnail", {
-								file_ids: selectedFiles.map((f) => f.id),
-								force: true,
-							});
-						},
-					},
-					{
-						icon: Sparkle,
-						label: "Generate Blurhashes",
-						onClick: async () => {
-							await runJob("thumbnail", {
-								file_ids: selectedFiles.map((f) => f.id),
-								generate_blurhash: true,
-							});
-						},
-						keybind: "⌘⇧B",
-					},
-					{
-						icon: TextAa,
-						label: "Extract Text from All",
-						onClick: async () => {
-							await runJob("ocr", {
-								file_ids: selectedFiles.map((f) => f.id),
-							});
-						},
-					},
-					{
-						icon: FilmStrip,
-						label: "Generate Thumbstrips (Videos)",
-						onClick: async () => {
-							const videos = selectedFiles.filter(
-								(f) => getContentKind(f) === "video",
-							);
-							if (videos.length > 0) {
-								await runJob("thumbstrip", {
-									file_ids: videos.map((f) => f.id),
-								});
-							}
-						},
-						condition: () =>
-							selectedFiles.some(
-								(f) => getContentKind(f) === "video",
-							),
-					},
-				],
-			},
-			{ type: "separator" },
-			{
-				icon: Trash,
-				label:
-					selected && selectedFiles.length > 1
-						? `Delete ${selectedFiles.length} items`
-						: "Delete",
-				onClick: async () => {
-					const targets = getTargetFiles();
-					const message =
-						targets.length > 1
-							? `Delete ${targets.length} items?`
-							: `Delete "${file.name}"?`;
-
-					if (confirm(message)) {
-						try {
-							await deleteFiles.mutateAsync({
-								targets: {
-									paths: targets.map((f) => f.sd_path),
-								},
-								permanent: false,
-								recursive: true,
-							});
-						} catch (err) {
-							console.error("Failed to delete:", err);
-						}
-					}
-				},
-				keybind: "⌘⌫",
-				variant: "danger" as const,
-			},
-		],
+	const contextMenu = useFileContextMenu({
+		file,
+		selectedFiles,
+		selected,
 	});
 
 	const handleClick = (e: React.MouseEvent) => {
